refactor: move router setup out of main into router module
