+++ frontend/src/pages/FinalizeTimetable.tsx
@@ -22,7 +22,7 @@ import { Button } from "../components/ui/button";
 import { Input } from "../components/ui/input";
 import { Label } from "../components/ui/label";
 import { useToast } from "../components/ui/use-toast";
-import { router } from "../main";
+import { router } from "../router";
 
 const finalizeTimetableRoute = new Route({
   getParentRoute: () => authenticatedRoute,
